requirements.txt
gunicorn==19.7.1
itsdangerous==0.24
Werkzeug==0.12.1
Pillow==4.2.1
Jinja2==2.9.6
Flask==0.12
Flask-RESTful==0.3.5
